requirements.txt
numpy>=1.24
scipy>=1.10
torch>=2.0
PyYAML>=6.0
pybullet>=3.2
